+++ Test/BLL/Customer/NewCustomersCompanyBL.cs
@@ -22,7 +22,7 @@ namespace Test.BLL.Customer
             this.newCustomersCompanyRepo = newCustomersCompanyRepo;
         }
 
-        public List<usp_NewCustomersCompany_Result> usp_NewCustomersCompany(DateTime start, DateTime end, int companyRef, int timeType)
+        public List<usp_NewCustomersCompany_Result> usp_NewCustomersCompany(DateTime start, DateTime end, long companyRef, int timeType)
         {
             return newCustomersCompanyRepo.usp_NewCustomersCompany(start, end, companyRef, timeType);
         }
