+++ Test/Models/Interface/Financial/IAvgDollarPerTransactionCompanyRepo.cs
@@ -9,6 +9,6 @@ namespace Test.Models.Interface.Financial
 {
     public interface IAvgDollarPerTransactionCompanyRepo: IDisposable
     {
-        List<usp_AvgDollarPerTransactionCompany_Result> usp_AvgDollarPerTransactionCompany(DateTime start, DateTime end, int companyRef, int timeType);
+        List<usp_AvgDollarPerTransactionCompany_Result> usp_AvgDollarPerTransactionCompany(DateTime start, DateTime end, long companyRef, int timeType);
     }
 }
