+++ Test/Models/Interface/Financial/IServiceOnlySalesCompanyRepo.cs
@@ -8,6 +8,6 @@ namespace Test.Models.Interface.Financial
 {
     public interface IServiceOnlySalesCompanyRepo : IDisposable
     {
-        List<usp_ServiceOnlySalesCompany_Result> usp_ServiceOnlySalesCompany(DateTime? start, DateTime? end, int companyRef, int timeType);
+        List<usp_ServiceOnlySalesCompany_Result> usp_ServiceOnlySalesCompany(DateTime? start, DateTime? end, long companyRef, int timeType);
     }
 }
